address review: dedupe dirname/filename injection, move tests out of regression dir
- Extract import_meta_dot and inject_dirname_filename_part helpers shared
  by the bundle-time and runtime lowering paths so declaration metadata
  stays identical in both
- Move the issue 4216 suite into test/bundler/cli.test.ts: the hardcoded
  path behavior was never correct, so it is not a regression
- Run --target=node bundles under a real node binary (skip if absent)
- Drain stdout/stderr on every build spawn
